update Todo type to Prisma generated type
The Todo type in this repo had an incorrect field: `created`.
We could easily fix this type, but it actually isn't the correct way to
use Prisma.

One of the main benefits of using Prisma is that the schema generates
types for you. If we import the Todo type from the Prisma client,
it will always stay up to date with the schema.